+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fixed issue where OAuth2 token was being renewed when no `refresh_token` was available.
+
 ## 0.23.1 - 2024-09-08
 
 - Fixed issue on auto-shutdown when specified `--parent-pid` dies (did not work if `robocorp.log` was not in the environment and it was not release dependency).